requirements.txt
torch>=1.13
torchvision
numpy
pandas
scikit-learn
matplotlib
tqdm
Pillow
opencv-python
scipy
seaborn
streamlit
torchcam
